+++ src/lib/builders/context-menu/index.ts
@@ -105,6 +105,7 @@ export function createContextMenu(args?: CreateContextMenu) {
 				}),
 				id: rootIds.menu,
 				'aria-labelledby': rootIds.trigger,
+				'data-melt-menu': '',
 				'data-state': $rootOpen ? 'open' : 'closed',
 				tabindex: -1,
 			} as const;
@@ -211,6 +212,7 @@ export function createContextMenu(args?: CreateContextMenu) {
 				'aria-expanded': $rootOpen,
 				'data-state': $rootOpen ? 'open' : 'closed',
 				id: rootIds.trigger,
+				'data-melt-menu-trigger': '',
 				style: styleToString({
 					WebkitTouchCallout: 'none',
 				}),