+++ src/lib/builders/menu/index.ts
@@ -66,6 +66,13 @@ export type CreateMenuArgs = {
 	 * @default true
 	 */
 	preventScroll?: boolean;
+
+	/**
+	 * Whether or not to loop the menu navigation.
+	 *
+	 * @default false
+	 */
+	loop?: boolean;
 };
 
 export type CreateSubmenuArgs = CreateMenuArgs & {
@@ -376,6 +383,7 @@ export function createMenuBuilder(opts: MenuBuilderOptions) {
 		}),
 		action: (node: HTMLElement, params: ItemArgs = {}) => {
 			const { onSelect } = params;
+			console.log('item', params);
 			setMeltMenuAttribute(node, selector);
 			applyAttrsIfDisabled(node);
 
@@ -1326,6 +1334,7 @@ function isMouse(e: PointerEvent) {
 export function setMeltMenuAttribute(element: HTMLElement | null, selector: Selector) {
 	if (!element) return;
 	const menuEl = element.closest(`${selector()}, ${selector('submenu')}`);
+	console.log('menueeeel', menuEl);
 
 	if (!isHTMLElement(menuEl)) return;
 	element.setAttribute('data-melt-menu-id', menuEl.id);
@@ -1340,14 +1349,17 @@ export function handleMenuNavigation(e: KeyboardEvent) {
 
 	// currently focused menu item
 	const currentFocusedItem = document.activeElement;
+	console.log(currentFocusedItem);
 	if (!isHTMLElement(currentFocusedItem)) return;
 
 	// menu element being navigated
 	const currentTarget = e.currentTarget;
+	console.log(currentTarget);
 	if (!isHTMLElement(currentTarget)) return;
 
 	// menu items of the current menu
 	const menuItems = getMenuItems(currentTarget);
+	console.log(menuItems);
 	if (!menuItems.length) return;
 
 	const candidateNodes = menuItems.filter((item) => {