+++ src/routes/(components)/h3.svelte
@@ -5,10 +5,6 @@
 	export { className as class };
 </script>
 
-<h3
-	class={cn('mb-1.5 mt-9 scroll-m-20 text-xl font-semibold', className)}
-	data-toc
-	{...$$restProps}
->
+<h3 class={cn('mt-9 scroll-m-20 text-xl font-semibold', className)} data-toc {...$$restProps}>
 	<slot />
 </h3>